scripts/requirements.txt
requests
anthropic
pyyaml
python-frontmatter
